requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.26.4
tensorflow==2.16.1
pyzbar==0.1.9
pillow==10.1.0